+++ edgar/cli/__init__.py
@@ -1,5 +1,8 @@
 # This directory is a Python package
 # Import all CLI modules for easy access
+from . import init
+from . import build
+from . import setup
 from . import add
 from . import new
 from . import probe
@@ -8,9 +11,7 @@ from . import select
 from . import update
 from . import report
 from . import calc
-from . import config
 from . import format
-from . import journal
 from . import modify
 from . import stats
 from . import shared